pyproject.toml
[tool.poetry]
name = "haistings"
version = "0.0.1"  # This will be overridden by poetry-dynamic-versioning
description = "AI-powered CVE prioritizer"
authors = ["Stacklok Inc"]
license = "Apache-2.0"
readme = "README.md"
packages = [
    { include = "haistings", from = "src" },
]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
langchain = {extras = ["openai"], version = "^0.3.18"}
langchain-community = "^0.3.17"
langchainhub = "^0.1.21"
langgraph = {extras = ["sqlite"], version = "^0.2.73"}
argparse = "^1.4.0"
gitingest = "^0.1.3"
gitpython = "^3.1.44"
langgraph-checkpoint-sqlite = "^2.0.5"
kubernetes = "^32.0.1"
ruff = "^0.9.9"
numpy = "^1.26.4"
sentence-transformers = "^2.6.0"
huggingface-hub = "^0.23.0"
langchain-huggingface = "^0.1.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
flake8 = "^7.0.0"
black = "^24.3.0"
isort = "^5.13.2"
mypy = "^1.9.0"
poetry-dynamic-versioning = "^1.2.0"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.poetry.scripts]
haistings = "haistings.haistings:main"

[tool.black]
line-length = 120
target-version = ["py312"]
include = ".pyi?$"

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "semver"
